requirements.txt
gradio>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
regex>=2024.5.15
